Biome id is encoded as byte for < 1.13

+++ src/protocolsupport/protocol/packet/middleimpl/clientbound/play/v_9r2_10_11_12r1_12r2/Chunk.java
@@ -41,7 +41,7 @@ public class Chunk extends AbstractChunk {
 			);
 			if (full) {
 				for (int i = 0; i < biomeData.length; i++) {
-					to.writeInt(biomeData[i]);
+					to.writeByte(biomeData[i]);
 				}
 			}
 		});
